# web/frontend/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_import_csv_roster_import_csv_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_import_csv_roster_import_csv_post",
        "type": "object"
      },
...
      "EmployeeIn": {
        "properties": {
          "can_cover_roles": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Can Cover Roles",
            "type": "array"
          },
          "id": {
            "pattern": "^[a-zA-Z0-9_-]{1,64}$",
            "title": "Id",
            "type": "string"
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Languages",
            "type": "array"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "nickname": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "phone": {
            "maxLength": 40,
            "minLength": 4,
            "title": "Phone",
            "type": "string"
          },
          "role": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Role",
            "type": "string"
          },
          "status": {
            "default": "active",
            "enum": [
              "active",
              "inactive",
              "terminated"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "role",
          "phone",
          "can_cover_roles"
        ],
        "title": "EmployeeIn",
        "type": "object"
      },
      "EmployeePatch": {
        "properties": {
          "can_cover_roles": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 20,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Can Cover Roles"
          },
          "languages": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 10,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Languages"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "nickname": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "phone": {
            "anyOf": [
              {
                "maxLength": 40,
                "minLength": 4,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "role": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "active",
                  "inactive",
                  "terminated"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "EmployeePatch",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SafetyToggleBody": {
        "properties": {
          "reason": {
            "maxLength": 200,
            "minLength": 5,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "SafetyToggleBody",
        "type": "object"
      },
      "ScheduleDayPut": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/ScheduleEntryIn"
            },
            "title": "Entries",
            "type": "array"
          }
        },
        "required": [
          "entries"
        ],
        "title": "ScheduleDayPut",
        "type": "object"
      },
      "ScheduleEntryIn": {
        "properties": {
          "employee_id": {
            "title": "Employee Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "shift": {
            "title": "Shift",
            "type": "string"
          }
        },
        "required": [
          "employee_id",
          "shift",
          "role"
        ],
        "title": "ScheduleEntryIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Shift Agent Cockpit API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/roster/employee": {
      "post": {
        "operationId": "add_employee_roster_employee_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Employee",
        "tags": [
          "roster"
        ]
      }
    },
    "/roster/employee/{employee_id}": {
      "delete": {
        "description": "Soft-delete: set status=terminated. Preserves history + phone for audits.",
        "operationId": "terminate_employee_roster_employee__employee_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "employee_id",
            "required": true,
            "schema": {
              "title": "Employee Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Terminate Employee",
        "tags": [
          "roster"
        ]
      },
      "patch": {
        "operationId": "patch_employee_roster_employee__employee_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "employee_id",
            "required": true,
            "schema": {
              "title": "Employee Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Employee",
        "tags": [
          "roster"
        ]
      }
    },
    "/roster/import-csv": {
      "post": {
        "description": "Atomic full-replace import of `roster.json` employees from CSV.\n\nRequired columns: id, name, role, phone, can_cover_roles\nOptional: nickname, languages, status\n\n`can_cover_roles` and `languages` are pipe-separated (`cashier|floor`)\nor comma-separated (`cashier,floor`) within their cell.\n\nRejects:\n- Cell values starting with =, +, -, @, \\t (formula injection)\n- Cells containing CR/LF (data corruption / smuggling)\n- Non-UTF-8 file content (with explicit 422, not 500)\n- File > 256 KB\n- Any row that fails Pydantic validation (Roster referential integrity\n  preserved by re-validating the FULL roster after replacement)",
        "operationId": "import_csv_roster_import_csv_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_csv_roster_import_csv_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Csv",
        "tags": [
          "roster"
        ]
      }
    },
    "/safety/disable": {
      "post": {
        "operationId": "disable_safety_disable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SafetyToggleBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable",
        "tags": [
          "safety"
        ]
      }
    },
    "/safety/enable": {
      "post": {
        "operationId": "enable_safety_enable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SafetyToggleBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enable",
        "tags": [
          "safety"
        ]
      }
    },
...
    "/schedule": {
      "get": {
        "operationId": "get_schedule_schedule_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "title": "To",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Schedule",
        "tags": [
          "schedule"
        ]
      }
    },
    "/schedule/{date}": {
      "delete": {
        "operationId": "delete_day_schedule__date__delete",
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "required": true,
            "schema": {
              "title": "Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Day",
        "tags": [
          "schedule"
        ]
      },
      "put": {
        "description": "Replace a day's full schedule (atomic).",
        "operationId": "put_day_schedule__date__put",
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "required": true,
            "schema": {
              "title": "Date",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleDayPut"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Day",
        "tags": [
          "schedule"
        ]
      }
    },
...
  }
}